+++ src/game/battlefield/mod.rs
@@ -4,6 +4,7 @@
 
 pub(in crate::game) mod components;
 pub(in crate::game) mod constants;
+pub(crate) mod ground_material;
 mod plugin;
 pub(in crate::game) mod systems;
 pub(in crate::game) mod trampling;